♻️ refactor: handleCheckboxClick을 TodoItem 내부에서 처리
