+++ src/features/TodoPage/TodoList.tsx
@@ -1,9 +1,4 @@
 import { DeleteTodoResponse, deleteTodo } from '@/apis/api/todos/deleteTodo';
-import {
-  UpdateTodoRequest,
-  UpdateTodoResponse,
-  putTodo,
-} from '@/apis/api/todos/putTodo';
 import { isAxiosErrorFromWantedPreOnboardingServer } from '@/apis/utils/isAxiosErrorFromWantedPreOnboardingServer';
 import { Todo } from '@/types/Todo';
 import { TodoItem } from './TodoItem';
@@ -14,40 +9,6 @@ type TodoListProps = {
 };
 
 export const TodoList = ({ todos, setTodos }: TodoListProps) => {
-  const handleCheckboxClick = async (
-    e: React.ChangeEvent<HTMLInputElement>,
-    todo: Todo,
-  ) => {
-    e.preventDefault();
-
-    const checked = e.target.checked;
-
-    const { id, todo: todoText } = todo;
-    const newUpdateTodoRequest: UpdateTodoRequest = {
-      todo: todoText,
-      isCompleted: checked,
-    };
-    try {
-      const res = await putTodo(id, newUpdateTodoRequest);
-      if (res.status === 200) {
-        const { id, todo: todoText, isCompleted } = res.data;
-        const newTodos = todos.map((todo) => {
-          if (todo.id === id) {
-            const newTodo: Todo = { id, todo: todoText, isCompleted };
-            return newTodo;
-          }
-          return todo;
-        });
-        setTodos(newTodos);
-      }
-    } catch (e: unknown) {
-      if (isAxiosErrorFromWantedPreOnboardingServer<UpdateTodoResponse>(e)) {
-        const { message } = e.response.data;
-        alert(message); // TODO: replace with toast
-      }
-    }
-  };
-
   const handleDeleteBtnClick = async (
     e: React.MouseEvent<HTMLButtonElement>,
     id: number,
@@ -85,7 +46,6 @@ export const TodoList = ({ todos, setTodos }: TodoListProps) => {
         <TodoItem
           key={todo.id}
           todo={todo}
-          onCheckboxClick={(e) => handleCheckboxClick(e, todo)}
           onDeleteBtnClick={(e) => handleDeleteBtnClick(e, todo.id)}
           setTodoInTodos={setTodoInTodos}
         />
